[#70] fix : 백엔드 배포 오류 수정작업
