+++ src/context/tag.context.tsx
@@ -8,7 +8,6 @@ import React, {
   createContext,
   PropsWithChildren,
   useCallback,
-  useContext,
   useEffect,
   useMemo,
   useRef,
@@ -28,15 +27,15 @@ interface TagContextType {
   handleClickDropDownData: (label: string) => void;
 }
 
-const TagContext = createContext<TagContextType | undefined>(undefined);
+export const TagContext = createContext<TagContextType | undefined>(undefined);
 
-export const useTagContext = () => {
-  const context = useContext(TagContext);
-  if (!context) {
-    throw new Error('TagContext 범위 밖입니다.');
-  }
-  return context;
-};
+// export const useTagContext = () => {
+//   const context = useContext(TagContext);
+//   if (!context) {
+//     throw new Error('TagContext 범위 밖입니다.');
+//   }
+//   return context;
+// };
 
 export const TagContextProvider = ({ children }: PropsWithChildren) => {
   const [tags, setTags] = useState<TagType[]>([]);
